Extract openapi configuration from app factory
