+++ poutine/robot_maker/templates/swagger-ui.html
@@ -40,7 +40,7 @@
 
     <script>
       window.onload = function() {
-        axios.get("/swagger-resources").then(function (response) {
+        axios.get("/swagger/resources").then(function (response) {
 
             // Build a system
             var ui = SwaggerUIBundle({
